Fix FTP verification.
It was only verifying connection, not logging on. (If you pass all three
of host, username and password to the initialiser it automatically logs
on, otherwise you need to call ``login()`` to accomplish that.)

The PASV bit was culled because it was bunkum; PASV only applies after
login (for data transfer) so it’s not relevant here.